Fix CRYPTO lot sizing to use Delta's real per-symbol contract_value; add CRYPTO symbol picker
Delta Exchange India's own get_lot_size() previously hardcoded `return 1`
for every CRYPTO perpetual future, when the real contract_value varies
per symbol (confirmed live against /v2/products: BTCUSD=0.001,
ETHUSD=0.01, DOGEUSD=100, etc.) - execution was sizing/rejecting CRYPTO
futures using 1 full underlying unit as the smallest tradeable size
instead of Delta's real per-lot size (a $555 available balance was
rejected as insufficient for BTCUSD when 1 real lot only costs ~$63.89).
Widens lot_size/get_lot_size from int to float across the three systems
that each hand-maintain a mirror of this shape (market-data, execution,
signal-generation - no shared code across systems/* by design), captures
contract_value during DeltaProvider.sync_instruments(), and fixes two
related bugs surfaced by this: an int() truncation in execution's own
HTTP client wrapper that silently zeroed every CRYPTO lot size and
crashed sizing with a ZeroDivisionError, and a balance-check ordering bug
in open_position/open_manual_position that compared available capital
against the cost of 1 full unit instead of 1 lot. Also fixes a float
round-trip bug in compute_risk_based_quantity's capital-cap recomputation
(caught by contract-guardian review) that silently under-counted by 1 lot
for many fractional-lot-size combinations.

The Manual tab's explicit "Lots" quantity field now actually multiplies
by the real lot_size too (open_manual_position previously used a typed
quantity as raw units directly, ignoring lot_size entirely) - and its
order-value preview does the same, with real currency labels (USD/INR)
so it doesn't misread as rupees by default.

Also adds GET /instruments/crypto-symbols (backs a proper CRYPTO symbol
dropdown on the Manual tab, sourced from Delta's own live product list,
instead of typing a symbol free-hand).

+++ systems/market-data/backend/app/domain/models.py
@@ -141,5 +141,5 @@ class ResolvedUnderlying(BaseModel):
     chart_exchange: str
     trade_symbol: str
     trade_exchange: str
-    lot_size: int
+    lot_size: float  # int for NSE/MCX F&O; a real fraction for Delta CRYPTO perpetuals (e.g. BTCUSD=0.001)
     expiry: Optional[str] = None  # ISO date - the trade contract's expiry
